Force max_reserve_timeout and timeout to be an int

+++ devices/__init__.py
@@ -100,9 +100,9 @@ class DefaultReserve(guacamole.Command):
                     logging.ERROR,
                     'Problem copying ssh key to device for: {}'.format(key))
         # default reservation timeout is 1 hour
-        timeout = reserve_data.get('timeout', '3600')
+        timeout = int(reserve_data.get('timeout', '3600'))
         # If max_reserve_timeout isn't specified, default to 18 hours
-        max_reserve_timeout = config.get('max_reserve_timeout', 18 * 60 * 60)
+        max_reserve_timeout = int(config.get('max_reserve_timeout', 18 * 60 * 60))
         if timeout > max_reserve_timeout:
             timeout = max_reserve_timeout
         print('*** TESTFLINGER SYSTEM RESERVED ***')
